Implement SVG Parser & Converter
Implemented the first version of the SVG parser and conversion module. It
can handle simple SVG documents, and will return a FNAME.vertex file
produced from the SVG file.

This commit includes a simple demo SVG document, as well as a partial
input2d file to demonstrate parsing abilities of the module.

This module has not yet been linked with the GUI.