Simplify File Loading and Write to Correct Folder
Simplified the loading of files by requiring only the SVG file to
be selected. The software then finds the corresponding input2d file
in the same folder.

The `outFile` has also been updated such that the `.vertex` files
will be written to the folder from which the SVG came. This will
simplify workflow for users.