Add CLI mode
Edited MeshmerizeMe to give the user the option of starting the software
either in GUI or CLI mode. The CLI mode was written to mirror the
actions taken in the `load_file` function of the GUI class, but printing
responses to the terminal instead of the GUI. The CLI `main` file
requires no user intervention.